+++ src/v2/templates/cart/cartempty.php
@@ -1,0 +1,27 @@
+<?php
+
+use ChurchCRM\dto\SystemURLs;
+
+require SystemURLs::getDocumentRoot() . '/Include/SimpleConfig.php';
+require SystemURLs::getDocumentRoot() . '/Include/Header.php';
+
+if (!array_key_exists('Message', $_GET)) {
+  ?>
+  <p class="text-center callout callout-warning"><?= gettext('You have no items in your cart.') ?> </p>
+  <?php
+} else {
+  switch ($_GET['Message']) {
+    case 'aMessage':
+      ?>
+      <p class="text-center callout callout-info"><?= $_GET['iCount'] . ' ' . ($_GET['iCount'] == 1 ? 'Record' : 'Records') . ' Emptied into Event ID:' . $_GET['iEID'] ?> </p>
+      <?php
+      break;
+  }
+}
+?>
+<p align="center">
+  <a href="<?= SystemURLs::getRootPath()?>/" class="btn btn-primary"><?=gettext('Back to Menu')?></a>
+</p>
+<?php
+
+require SystemURLs::getDocumentRoot() . '/Include/Footer.php';
